requirements.txt
dj-database-url==0.4.2
Django==1.11.2
django-mathfilters==0.4.0
djangorestframework==3.7.7
gunicorn==19.7.1
psycopg2==2.7.1
pytz==2017.3
simplejson==3.11.1
whitenoise==3.3.1
yahoo-finance==1.4.0